requirements.txt
asgiref==3.2.10
Django==3.1
django-nested-inline==0.4.2
djangorestframework==3.11.1
psycopg2==2.8.5
python-decouple==3.3
pytz==2020.1
sqlparse==0.3.1
